vrrpd: fix unsigned - signed cmp
For some reason this warning only shows up on armel

+++ vrrpd/vrrp_vty.c
@@ -555,7 +555,7 @@ DEFPY(vrrp_vrid_show,
 	for (ALL_LIST_ELEMENTS_RO(ll, ln, vr)) {
 		if (ifn && !strmatch(ifn, vr->ifp->name))
 			continue;
-		if (vrid && vrid != vr->vrid)
+		if ((uint8_t) vrid && vrid != vr->vrid)
 			continue;
 
 		if (!json)
